refactor: move package into src
