+++ pyproject.toml
@@ -1,2 +1,3 @@
 [build-system]
-requires = ['setuptools>=30.3.0', 'wheel', 'numpy', 'setuptools_scm']
+requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]>=3.4", "numpy"]
+build-backend = "setuptools.build_meta"
