+++ src/read_acq/codec.py
