+++ tests/test_encoding.py
@@ -1,6 +1,5 @@
-from read_acq.codec import _encode_line, _decode_line, _encode
-from read_acq import encode
-from io import StringIO
+from src.read_acq.codec import _encode_line, _decode_line, _encode
+from src.read_acq import encode
 
 import numpy as np
